requirements.txt
# Core dependencies
fastapi_poe==0.0.36
feedparser==6.0.11
PyGithub==2.1.1
Jinja2==3.1.3
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0

# LangGraph & LangChain stack
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community==0.3.5

# LlamaIndex & vector storage
llama-index>=0.11.0
chromadb>=0.5.0
sentence-transformers>=2.2.0

# Topic modeling & clustering
bertopic>=0.16.0
hdbscan>=0.8.33
umap-learn>=0.5.5
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.4.2

# Statistics / testing
scipy>=1.11.0

# AWS SDK (for deployment)
boto3==1.34.34

# Date handling
python-dateutil==2.8.2

# HTML parsing (for some RSS feeds)
beautifulsoup4==4.12.3
lxml==5.1.0

# Reddit API
praw==7.7.1

